Add parameters to JsCode
